sqrt->sqrtf

+++ quaternonFilters.ino
@@ -39,7 +39,7 @@
             float q4q4 = q4 * q4;
 
             // Normalise accelerometer measurement
-            norm = sqrt(ax * ax + ay * ay + az * az);
+            norm = sqrtf(ax * ax + ay * ay + az * az);
             if (norm == 0.0f) return; // handle NaN
             norm = 1.0f/norm;
             ax *= norm;
@@ -47,7 +47,7 @@
             az *= norm;
 
             // Normalise magnetometer measurement
-            norm = sqrt(mx * mx + my * my + mz * mz);
+            norm = sqrtf(mx * mx + my * my + mz * mz);
             if (norm == 0.0f) return; // handle NaN
             norm = 1.0f/norm;
             mx *= norm;
@@ -71,7 +71,7 @@
             s2 = _2q4 * (2.0f * q2q4 - _2q1q3 - ax) + _2q1 * (2.0f * q1q2 + _2q3q4 - ay) - 4.0f * q2 * (1.0f - 2.0f * q2q2 - 2.0f * q3q3 - az) + _2bz * q4 * (_2bx * (0.5f - q3q3 - q4q4) + _2bz * (q2q4 - q1q3) - mx) + (_2bx * q3 + _2bz * q1) * (_2bx * (q2q3 - q1q4) + _2bz * (q1q2 + q3q4) - my) + (_2bx * q4 - _4bz * q2) * (_2bx * (q1q3 + q2q4) + _2bz * (0.5f - q2q2 - q3q3) - mz);
             s3 = -_2q1 * (2.0f * q2q4 - _2q1q3 - ax) + _2q4 * (2.0f * q1q2 + _2q3q4 - ay) - 4.0f * q3 * (1.0f - 2.0f * q2q2 - 2.0f * q3q3 - az) + (-_4bx * q3 - _2bz * q1) * (_2bx * (0.5f - q3q3 - q4q4) + _2bz * (q2q4 - q1q3) - mx) + (_2bx * q2 + _2bz * q4) * (_2bx * (q2q3 - q1q4) + _2bz * (q1q2 + q3q4) - my) + (_2bx * q1 - _4bz * q3) * (_2bx * (q1q3 + q2q4) + _2bz * (0.5f - q2q2 - q3q3) - mz);
             s4 = _2q2 * (2.0f * q2q4 - _2q1q3 - ax) + _2q3 * (2.0f * q1q2 + _2q3q4 - ay) + (-_4bx * q4 + _2bz * q2) * (_2bx * (0.5f - q3q3 - q4q4) + _2bz * (q2q4 - q1q3) - mx) + (-_2bx * q1 + _2bz * q3) * (_2bx * (q2q3 - q1q4) + _2bz * (q1q2 + q3q4) - my) + _2bx * q2 * (_2bx * (q1q3 + q2q4) + _2bz * (0.5f - q2q2 - q3q3) - mz);
-            norm = sqrt(s1 * s1 + s2 * s2 + s3 * s3 + s4 * s4);    // normalise step magnitude
+            norm = sqrtf(s1 * s1 + s2 * s2 + s3 * s3 + s4 * s4);    // normalise step magnitude
             norm = 1.0f/norm;
             s1 *= norm;
             s2 *= norm;
@@ -89,7 +89,7 @@
             q2 += qDot2 * deltat;
             q3 += qDot3 * deltat;
             q4 += qDot4 * deltat;
-            norm = sqrt(q1 * q1 + q2 * q2 + q3 * q3 + q4 * q4);    // normalise quaternion
+            norm = sqrtf(q1 * q1 + q2 * q2 + q3 * q3 + q4 * q4);    // normalise quaternion
             norm = 1.0f/norm;
             q[0] = q1 * norm;
             q[1] = q2 * norm;
@@ -124,7 +124,7 @@
             float q4q4 = q4 * q4;   
 
             // Normalise accelerometer measurement
-            norm = sqrt(ax * ax + ay * ay + az * az);
+            norm = sqrtf(ax * ax + ay * ay + az * az);
             if (norm == 0.0f) return; // handle NaN
             norm = 1.0f / norm;        // use reciprocal for division
             ax *= norm;
@@ -132,7 +132,7 @@
             az *= norm;
 
             // Normalise magnetometer measurement
-            norm = sqrt(mx * mx + my * my + mz * mz);
+            norm = sqrtf(mx * mx + my * my + mz * mz);
             if (norm == 0.0f) return; // handle NaN
             norm = 1.0f / norm;        // use reciprocal for division
             mx *= norm;
@@ -142,7 +142,7 @@
             // Reference direction of Earth's magnetic field
             hx = 2.0f * mx * (0.5f - q3q3 - q4q4) + 2.0f * my * (q2q3 - q1q4) + 2.0f * mz * (q2q4 + q1q3);
             hy = 2.0f * mx * (q2q3 + q1q4) + 2.0f * my * (0.5f - q2q2 - q4q4) + 2.0f * mz * (q3q4 - q1q2);
-            bx = sqrt((hx * hx) + (hy * hy));
+            bx = sqrtf((hx * hx) + (hy * hy));
             bz = 2.0f * mx * (q2q4 - q1q3) + 2.0f * my * (q3q4 + q1q2) + 2.0f * mz * (0.5f - q2q2 - q3q3);
 
             // Estimated direction of gravity and magnetic field
@@ -185,7 +185,7 @@
             q4 = pc + (q1 * gz + pa * gy - pb * gx) * (0.5f * deltat);
 
             // Normalise quaternion
-            norm = sqrt(q1 * q1 + q2 * q2 + q3 * q3 + q4 * q4);
+            norm = sqrtf(q1 * q1 + q2 * q2 + q3 * q3 + q4 * q4);
             norm = 1.0f / norm;
             q[0] = q1 * norm;
             q[1] = q2 * norm;
